+++ crates/matrix-sdk-base/src/room/room_info.rs
@@ -50,7 +50,7 @@ use ruma::{
     OwnedRoomId, OwnedUserId, RoomAliasId, RoomId, RoomVersionId, UserId,
 };
 use serde::{Deserialize, Serialize};
-use tracing::{debug, field::debug, info, instrument, warn};
+use tracing::{debug, error, field::debug, info, instrument, warn};
 
 use super::{
     AccountDataSource, EncryptionState, Room, RoomCreateWithCreatorEventContent, RoomDisplayName,
@@ -65,6 +65,18 @@ use crate::{
     sync::UnreadNotificationsCount,
     MinimalStateEvent, OriginalMinimalStateEvent,
 };
+
+/// A struct remembering details of an invite and if the invite has been
+/// accepted on this particular client.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct InviteAcceptanceDetails {
+    /// A timestamp remembering when we observed the user accepting an invite
+    /// using this client.
+    pub invite_accepted_at: MilliSecondsSinceUnixEpoch,
+
+    /// The user ID of the person that invited us.
+    pub inviter: OwnedUserId,
+}
 
 impl Room {
     /// Subscribe to the inner `RoomInfo`.
@@ -471,7 +483,7 @@ pub struct RoomInfo {
     /// This is useful to remember if the user accepted this a join on this
     /// specific client.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) invite_accepted_at: Option<MilliSecondsSinceUnixEpoch>,
+    pub(crate) invite_acceptance_details: Option<InviteAcceptanceDetails>,
 }
 
 impl RoomInfo {
@@ -494,7 +506,7 @@ impl RoomInfo {
             cached_display_name: None,
             cached_user_defined_notification_mode: None,
             recency_stamp: None,
-            invite_accepted_at: None,
+            invite_acceptance_details: None,
         }
     }
 
@@ -525,6 +537,12 @@ impl RoomInfo {
 
     /// Set the membership RoomState of this Room
     pub fn set_state(&mut self, room_state: RoomState) {
+        if self.state() != RoomState::Joined && self.invite_acceptance_details.is_some() {
+            error!(room_id = %self.room_id, "The RoomInfo contains invite acceptance details but the room is not in the joined state");
+        }
+        // Changing our state removes the invite details since we can't know that they
+        // are relevant anymore.
+        self.invite_acceptance_details = None;
         self.room_state = room_state;
     }
 
@@ -758,10 +776,8 @@ impl RoomInfo {
         self.summary.invited_member_count = count;
     }
 
-    /// Mark that the user has accepted an invite and remember when this has
-    /// happened using a timestamp set to [`MilliSecondsSinceUnixEpoch::now()`].
-    pub(crate) fn set_invite_accepted_now(&mut self) {
-        self.invite_accepted_at = Some(MilliSecondsSinceUnixEpoch::now());
+    pub(crate) fn set_invite_acceptance_details(&mut self, details: InviteAcceptanceDetails) {
+        self.invite_acceptance_details = Some(details);
     }
 
     /// Returns the timestamp when an invite to this room has been accepted by
@@ -770,8 +786,8 @@ impl RoomInfo {
     /// # Returns
     /// - `Some` if the invite has been accepted by this specific client.
     /// - `None` if the invite has not been accepted
-    pub fn invite_accepted_at(&self) -> Option<MilliSecondsSinceUnixEpoch> {
-        self.invite_accepted_at
+    pub fn invite_acceptance_details(&self) -> Option<InviteAcceptanceDetails> {
+        self.invite_acceptance_details.clone()
     }
 
     /// Updates the room heroes.
@@ -1247,7 +1263,7 @@ mod tests {
             cached_display_name: None,
             cached_user_defined_notification_mode: None,
             recency_stamp: Some(42),
-            invite_accepted_at: None,
+            invite_acceptance_details: None,
         };
 
         let info_json = json!({